+++ extensions/CuteHMI/Modbus.4/HoldingRegisterItem.qml
@@ -28,7 +28,7 @@ Item {
 
 		parent: root
 		visible: root.enabled
-		running: root.controller.busy
+		running: root.controller.busy || !controller.initialized
 	}
 
 	// Private properties.
@@ -38,14 +38,25 @@ Item {
 		property var delegateValue: delegate[delegateProperty]
 	}
 
+	Connections {
+		target: controller
+
+		function onInitializedChanged() {
+			if (controller.initialized)
+				Private.onInitialized()
+			else
+				Private.disconnect()
+		}
+	}
+
 	Component.onCompleted: Private.onCompleted()
 
-	Component.onDestruction: Private.onDestruction()
+	Component.onDestruction: Private.disconnect()
 
 	onReadOnlyChanged: Private.onReadOnlyChanged()
 }
 
-//(c)C: Copyright © 2022-2023, Michał Policht <michal@policht.pl>. All rights reserved.
+//(c)C: Copyright © 2022-2024, Michał Policht <michal@policht.pl>. All rights reserved.
 //(c)C: SPDX-License-Identifier: LGPL-3.0-or-later OR MIT
 //(c)C: This file is a part of CuteHMI.
 //(c)C: CuteHMI is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.