+++ extensions/CuteHMI/Modbus.4/RegisterItem.js
@@ -1,12 +1,20 @@
 function onCompleted()
 {
+	if (controller.initialized)
+		onInitialized()
+	else
+		delegate[delegateProperty] = controller.value
+}
+
+function onInitialized()
+{
 	delegate[delegateProperty] = controller.value
 	controller.valueUpdated.connect(onValueUpdatedFromBackend)
 	if (!readOnly)
 		p.delegateValueChanged.connect(onValueChangedFromFrontend)
 }
 
-function onDestruction()
+function disconnect()
 {
 	controller.valueUpdated.disconnect(onValueUpdatedFromBackend)
 	p.delegateValueChanged.disconnect(onValueChangedFromFrontend)
@@ -36,7 +44,7 @@ function onValueChangedFromFrontend()
 	controller.value = p.delegateValue
 }
 
-//(c)C: Copyright © 2022, Michał Policht <michal@policht.pl>. All rights reserved.
+//(c)C: Copyright © 2022-2024, Michał Policht <michal@policht.pl>. All rights reserved.
 //(c)C: SPDX-License-Identifier: LGPL-3.0-or-later OR MIT
 //(c)C: This file is a part of CuteHMI.
 //(c)C: CuteHMI is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.