+++ extensions/CuteHMI/Modbus.4/src/cutehmi/modbus/Register1Controller.cpp
@@ -97,6 +97,7 @@ void Register1Controller::onRequestCompleted(QJsonObject request, QJsonObject re
 
 void Register1Controller::resetRegister()
 {
+	setInitialized(false);
 	m->requestId = QUuid();	// Setting up new register invalidates previous requests.
 	m->postponedWritePending = false;
 	m->adjustingValue = false;
@@ -106,6 +107,7 @@ void Register1Controller::resetRegister()
 
 	if (device()) {
 		m->register1 = registerAt(static_cast<quint16>(address()));
+		updateValue(m->register1->value());
 		if (enabled()) {
 			setBusy(true);
 			m->register1->awake();
@@ -132,7 +134,7 @@ bool Register1Controller::verifyRegisterValue() const
 }
 }
 
-//(c)C: Copyright © 2022-2023, Michał Policht <michal@policht.pl>. All rights reserved.
+//(c)C: Copyright © 2022-2024, Michał Policht <michal@policht.pl>. All rights reserved.
 //(c)C: SPDX-License-Identifier: LGPL-3.0-or-later OR MIT
 //(c)C: This file is a part of CuteHMI.
 //(c)C: CuteHMI is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.