+++ extensions/CuteHMI/Modbus.4/CHANGES.md
@@ -3,6 +3,8 @@
 ## Version 4
 
 - This version has switched from CuteHMI.Services.2 to CuteHMI.Services.3.
+- Property `initialized` has been introduced in cutehmi::modbus::AbstractRegisterController. Controller items use it for more robust
+delegate initialization.
 
 ## Version 3